+++ src/hooks/product/index.ts
@@ -1,0 +1,5 @@
+export { default as useProductBasicInfo } from './useProductBasicInfo';
+export { default as useProductDetailInfo } from './useProductDetailInfo';
+export { default as useProductReview } from './useProductReview';
+export { default as useProductWish } from './useProductWish';
+export { default as useProductDetail } from './useProductDetail';
