+++ src/pages/ProductDetail/ProductDetailFooter.tsx
@@ -3,7 +3,7 @@ import {
   FooterLikeButton,
   FooterOrderButton,
 } from '@/styles/Product/ProductFooter.styles';
-import useProductDetail from '@/hooks/useProductDetail';
+import useProductDetail from '@/hooks/product/useProductDetail';
 import { HiHeart, HiOutlineHeart } from 'react-icons/hi';
 
 function ProductDetailFooter() {
